// app/interfaces/MainComponent/ProtectSection.ts
export interface IProtectSection {
    _id: string;
    id?: number;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}

export interface IProtectSectionInput {
    id?: number;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}